Logo refinment and small lint fixes
