+++ hat/assets/js/apps/Iaso/domains/app/components/Logo.tsx
@@ -1,48 +1,19 @@
-import React, { FunctionComponent, useContext } from 'react';
-import { Box, Theme } from '@mui/material';
-
-import { ThemeConfigContext } from '../contexts/ThemeConfigContext';
+import React, { FC } from 'react';
 import { LogoSvg } from './LogoSvg';
 
-const styles = {
-    text: {
-        fontFamily: '"DINAlternate-Bold", "DIN Alternate", sans-serif',
-        fontSize: 23,
-    },
+type Props = {
+    isIaso: boolean;
+    logoPath: string;
 };
 
-export const Logo: FunctionComponent = () => {
-    const { LOGO_PATH, APP_TITLE, SHOW_NAME_WITH_LOGO } =
-        useContext(ThemeConfigContext);
-    const showAppName = SHOW_NAME_WITH_LOGO === 'yes';
-
-    let logoComponent: React.ReactNode = null;
-    if (APP_TITLE === 'Iaso') {
-        logoComponent = <LogoSvg />;
-    } else if (LOGO_PATH) {
-        logoComponent = (
-            <img
-                alt="logo"
-                src={`${window.STATIC_URL}${LOGO_PATH}`}
-                style={{ maxHeight: '50px', maxWidth: '200px' }}
-            />
-        );
-    }
-
-    return (
-        <>
-            {logoComponent}
-            {showAppName && (
-                <Box
-                    component="span"
-                    sx={(theme: Theme) => ({
-                        ...styles.text,
-                        marginLeft: logoComponent ? theme.spacing(2) : 0,
-                    })}
-                >
-                    {APP_TITLE}
-                </Box>
-            )}
-        </>
+export const Logo: FC<Props> = ({ isIaso = true, logoPath = '' }) => {
+    return isIaso ? (
+        <LogoSvg />
+    ) : (
+        <img
+            alt="logo"
+            src={`${window.STATIC_URL}${logoPath}`}
+            style={{ maxHeight: '50px', maxWidth: '200px' }}
+        />
     );
 };
